+++ src/blas_axpy.cpp
@@ -19,7 +19,7 @@ do_setup(const benchmark::State & state)
     x = new double[n];
     y = new double[n];
 
-    for (i32 i = 0; i < n; ++i) {
+    for (i64 i = 0; i < n; ++i) {
         x[i] = std::rand();
         y[i] = std::rand();
     }
