+++ src/blas_scale.cpp
@@ -17,7 +17,7 @@ do_setup(const benchmark::State & state)
     auto n = state.range(0);
     y = new double[n];
 
-    for (i32 i = 0; i < n; ++i) {
+    for (i64 i = 0; i < n; ++i) {
         y[i] = std::rand();
     }
 }
